// src/components/Sidebar/menu/types.ts
import { type PermissionEnum } from "@dashboard/graphql";
import { type Sprinkles } from "@saleor/macaw-ui-next";
import { type ReactNode } from "react";

export interface SidebarMenuItem {
  label?: string | ReactNode;
  id: string;
  url?: string;
  matchUrls?: string[];
  permissions?: PermissionEnum[];
  requireAllPermissions?: boolean;
  hidden?: boolean;
  type: "item" | "itemGroup" | "divider";
  icon?: ReactNode;
  onClick?: () => void;
  children?: SidebarMenuItem[];
  paddingY?: Sprinkles["paddingY"];
  endAdornment?: ReactNode;
}